Address PR #18 feedback: remove comment bloat, delete unused Delete methods, refactor tests
- Revert excessive comment commit (35e6396)
- Add WithTx doc comment clarifying callers must use txStore repos
- Fix TestMain to close store before log.Fatalf on pool failure
- Remove DeleteFrom/DeleteFromBlock from interfaces, implementations, fakes, and tests
- Split monolithic postgres_test.go into per-repo test files with shared helpers
- Use table-driven tests and cmp.Diff for struct comparisons
