+++ .gitignore
@@ -1,1 +1,2 @@
 llm_ignore
+.claude/scheduled_tasks.lock
