+++ internal/store/postgres/block.go
@@ -38,12 +38,3 @@ func (r *BlockRepo) GetHash(ctx context.Context, chainID int64, blockNumber uint
 	}
 	return hash, nil
 }
-
-// DeleteFrom removes all blocks for the given chain at or above fromBlock.
-func (r *BlockRepo) DeleteFrom(ctx context.Context, chainID int64, fromBlock uint64) error {
-	_, err := r.db.Exec(ctx,
-		"DELETE FROM blocks WHERE chain_id = $1 AND block_number >= $2",
-		chainID, fromBlock,
-	)
-	return err
-}
